# backend/openapi/TidySense.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TidySense | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageDtoOfProjectDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageDtoOfProjectDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPageDtoOfProjectDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}/review": {
      "post": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewProjectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewProjectRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}/terminal-preview": {
      "post": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalPreviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalPreviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalPreviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalPreviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalPreviewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{id}/terminal": {
      "post": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalCommandRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalCommandRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiProblemDto": {
        "required": [
          "type",
          "title",
          "status",
          "code",
          "messageKey",
          "retryable",
          "traceId",
          "correlationId",
          "detail",
          "errors",
          "entityId",
          "expectedVersion",
          "currentVersion"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "messageKey": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "traceId": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "entityId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "currentVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "CreateProjectRequest": {
        "required": [
          "title",
          "completionMeaning",
          "goalId",
          "targetDate",
          "reviewDate"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "completionMeaning": {
            "type": [
              "null",
              "string"
            ]
          },
          "goalId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "targetDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "reviewDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          }
        }
      },
...
      "CursorPageDtoOfProjectDto": {
        "required": [
          "items",
          "page"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectDto"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageInfoDto"
          }
        }
      },
...
      "PageInfoDto": {
        "required": [
          "nextCursor",
          "hasMore"
        ],
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "ProjectDto": {
        "required": [
          "id",
          "goalId",
          "title",
          "completionMeaning",
          "status",
          "targetDate",
          "reviewDate",
          "reviewDateSource",
          "source",
          "version",
          "createdAt",
          "updatedAt",
          "terminalAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "goalId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "completionMeaning": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "targetDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "reviewDate": {
            "type": "string",
            "format": "date"
          },
          "reviewDateSource": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "terminalAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ReviewProjectRequest": {
        "required": [
          "reviewDate",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "reviewDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "TerminalBlockerDto": {
        "required": [
          "resourceType",
          "resourceId",
          "status",
          "version"
        ],
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string"
          },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "TerminalCascadeDto": {
        "required": [
          "resourceType",
          "resourceId",
          "resultingStatus",
          "version"
        ],
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string"
          },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "resultingStatus": {
            "type": "string"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "TerminalCommandRequest": {
        "required": [
          "targetStatus",
          "expectedVersion",
          "previewHash"
        ],
        "type": "object",
        "properties": {
          "targetStatus": {
            "type": "string"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "previewHash": {
            "type": "string"
          }
        }
      },
      "TerminalPreviewDto": {
        "required": [
          "entityId",
          "entityType",
          "currentStatus",
          "targetStatus",
          "expectedVersion",
          "canApply",
          "blockers",
          "cascades",
          "previewHash"
        ],
        "type": "object",
        "properties": {
          "entityId": {
            "type": "string",
            "format": "uuid"
          },
          "entityType": {
            "type": "string"
          },
          "currentStatus": {
            "type": "string"
          },
          "targetStatus": {
            "type": "string"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "canApply": {
            "type": "boolean"
          },
          "blockers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TerminalBlockerDto"
            }
          },
          "cascades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TerminalCascadeDto"
            }
          },
          "previewHash": {
            "type": "string"
          }
        }
      },
      "TerminalPreviewRequest": {
        "required": [
          "targetStatus",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "targetStatus": {
            "type": "string"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "UpdateProjectRequest": {
        "required": [
          "title",
          "completionMeaning",
          "goalId",
          "targetDate",
          "reviewDate",
          "expectedVersion"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "completionMeaning": {
            "type": [
              "null",
              "string"
            ]
          },
          "goalId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "targetDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "reviewDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "expectedVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}